+++ app/Http/Controllers/UsersController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\User;
+use App\Models\Data\User;
 use Illuminate\Http\Request;
 
 class UsersController extends Controller
