+++ app/Models/Data/Contributor.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models\Data;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,6 +8,6 @@ class Contributor extends Model
 {
     public function type()
     {
-        return $this->belongsTo('App\ContributorType');
+        return $this->belongsTo('App\Models\Data\ContributorType');
     }
 }
