+++ app/Models/Data/CreatureType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models\Data;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -10,6 +10,6 @@ class CreatureType extends Model
 
     public function creatures()
     {
-        return $this->hasMany('App\Creature');
+        return $this->hasMany('App\Models\Data\Creature');
     }
 }
