+++ app/Models/Data/Item.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App;
+namespace App\Models\Data;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Publisher extends Model
+class Item extends Model
 {
     //
 }
