+++ app/Models/Data/ModuleRating.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models\Data;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,11 +8,11 @@ class ModuleRating extends Model
 {
     public function user()
     {
-        return $this->belongsTo('App\User');
+        return $this->belongsTo('App\Models\Data\User');
     }
 
     public function module()
     {
-        return $this->belongsTo('App\Module');
+        return $this->belongsTo('App\Models\Data\Module');
     }
 }
